Add `test_densities.py` & `test_plasma_critical_density()` Add `densities.py` & move `plasma_critical_density()` from `frequencies.py` to `densities.py` Revise changelog entry
